requirements.txt
Flask
requests
python-dotenv
sec-parser
sec-edgar-downloader
matplotlib
numpy
argparse